Use local line splitter
