+++ test-env/src/main/kotlin/io/spine/protodata/test/AnnotationInsertionPointPrinter.kt
@@ -29,6 +29,7 @@ package io.spine.protodata.test
 import io.spine.protodata.TextCoordinates
 import io.spine.protodata.renderer.InsertionPoint
 import io.spine.protodata.renderer.InsertionPointPrinter
+import io.spine.protodata.splitLines
 import io.spine.text.Text
 import io.spine.tools.code.CommonLanguages.Java
 import kotlin.text.RegexOption.DOT_MATCHES_ALL
@@ -44,7 +45,7 @@ public enum class AnnotationInsertionPoint : InsertionPoint {
 
     IMPORT {
         override fun locate(text: Text): TextCoordinates {
-            val lines = text.lines()
+            val lines = text.splitLines()
             val packageLineIndex = lines.asSequence()
                 .mapIndexed { index, line -> index to line }
                 .find { (_, line) -> line.startsWith("package") }
@@ -55,7 +56,7 @@ public enum class AnnotationInsertionPoint : InsertionPoint {
     },
     BEFORE_RETURN_TYPE_METHOD_FOO {
         override fun locate(text: Text): TextCoordinates {
-            val lines = text.lines()
+            val lines = text.splitLines()
             val (lineIndex, line) = lines.asSequence()
                 .mapIndexed { index, line -> index to line }
                 .find { (_, line) ->
